// electro-pime/types/api.ts
export interface Product {
  id: string;
  name: string;
  description?: string;
  price: number;
  cost: number;
  stock: number;
  category: string;
  isActive: boolean;
  imageUrl?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductDto {
  name: string;
  description?: string;
  price: number;
  cost: number;
  stock: number;
  category: string;
  isActive?: boolean;
  imageUrl?: string;
}